Restructure library_fetch so it can be run from a worker.
